STARRY: PSST-7312: Allow callers to return application level errors
(Keep in mind when reading bluez related comments below that this
pertains to bluez 5.72)

When we return an error from a read or write response, if we pack an
error code into the string returned to `BLUEZ_ERROR_FAILED` (which is
"org.bluez.Error.Failed"), that ends up being returned to the BT caller.
Previously, we returned a blank string which is always interpreted as an
"Unlikely error" (code 0x0e) by bluez. This was suboptimal
since it could potentially make the caller think that something is very wrong
in some lower level in the BT stack when in reality the server received and
processed the request and sent back a response.

For posterity, we can also return other specific error codes by
returning to different dbus locations. For example:

    MethodError::from(("org.bluez.Error.NotPermitted", ""))

ends up returning either error code 0x02 (Read not
permitted) or 0x03 (Write not permitted) to the caller, depending on
whether a read or a write was attempted.

See this bluez code for more info:
https://github.com/bluez/bluez/blob/bd7d49d54aa3aa490ebdd67b3dd2317d29213d45/src/gatt-database.c#L2216

Also see the BT core spec here:
https://www.bluetooth.com/specifications/specs/core-specification-5-3/